+++ packages/stream_chat_flutter/example/windows/flutter/generated_plugins.cmake
@@ -9,11 +9,11 @@ list(APPEND FLUTTER_PLUGIN_LIST
   gal
   media_kit_video
   record_windows
-  screen_brightness_windows
   share_plus
   sqlite3_flutter_libs
   thumblr_windows
   url_launcher_windows
+  volume_controller
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST