+++ packages/stream_chat_flutter/test/src/attachment/file_attachment_test.dart
@@ -44,6 +44,9 @@ void main() {
         ),
       );
 
+      // wait for the initial state to be rendered.
+      await tester.pumpAndSettle();
+
       expect(find.text('example.pdf'), findsOneWidget);
     },
   );