+++ packages/stream_chat_flutter/test/src/attachment/url_attachment_test.dart
@@ -39,6 +39,9 @@ void main() {
         ),
       );
 
+      // wait for the initial state to be rendered.
+      await tester.pumpAndSettle();
+
       expect(find.text('Flutter'), findsOneWidget);
     },
   );