+++ packages/stream_chat_flutter/test/src/channel/channel_name_test.dart
@@ -72,6 +72,9 @@ void main() {
         ),
       );
 
+      // wait for the initial state to be rendered.
+      await tester.pumpAndSettle();
+
       expect(find.text('test'), findsOneWidget);
     },
   );