+++ packages/stream_chat_flutter/test/src/dialogs/channel_info_dialog_test.dart
@@ -36,7 +36,6 @@ void main() {
       ),
     ]);
     when(() => channel.name).thenReturn('test-channel');
-    when(() => channel.id).thenReturn('123456789');
     when(() => channel.isDistinct).thenReturn(true);
     when(() => channel.memberCount).thenReturn(2);
     when(() => channelState.membersStream).thenAnswer(
@@ -71,6 +70,9 @@ void main() {
         ),
       ),
     );
+
+    // wait for the initial state to be rendered.
+    await tester.pump(Duration.zero);
 
     expect(find.byType(SimpleDialog), findsOneWidget);
     expect(find.byType(StreamChannelInfo), findsOneWidget);