+++ packages/stream_chat_flutter/test/src/indicators/typing_indicator_test.dart
@@ -86,6 +86,9 @@ void main() {
         ),
       ));
 
+      // wait for the initial state to be rendered.
+      await tester.pump(Duration.zero);
+
       expect(find.byKey(typingKey), findsOneWidget);
       expect(find.byType(Flexible), findsOneWidget);
     },