+++ packages/stream_chat_flutter/test/src/message_list_view/bottom_row_test.dart
@@ -63,6 +63,9 @@ void main() {
       ),
     );
 
+    // wait for the initial state to be rendered.
+    await tester.pump(Duration.zero);
+
     await tester.tap(find.byType(GestureDetector));
     await tester.pumpAndSettle();