+++ packages/stream_chat_flutter_core/test/mocks.dart
@@ -29,16 +29,21 @@ class MockClientState extends Mock implements ClientState {
       );
 }
 
-class MockChannel extends Mock implements Channel {
-  ChannelClientState? _state;
-
-  @override
-  ChannelClientState get state => _state ??= MockChannelState();
-
+class NonInitializedMockChannel extends Mock implements Channel {
   StreamChatClient? _client;
 
   @override
   StreamChatClient get client => _client ??= MockClient();
+
+  @override
+  ChannelClientState? get state => null;
+}
+
+class MockChannel extends NonInitializedMockChannel {
+  ChannelClientState? _state;
+
+  @override
+  ChannelClientState get state => _state ??= MockChannelState();
 }
 
 class MockChannelState extends Mock implements ChannelClientState {}